Add files via upload
Hello, I've made an front end basic design using HTML, CSS & Bootstrap. As a practice, I've made this resume to to refine & improve my skills. It contains different elements like images, sections, containers, flex, etc.
Disclaimer: The content mentioned in the resume is not 100% legit, consider it as a one of the design templates.